Refactor names to reflect eng sync discussion
