+++ ozaria/engine/cinematic/Command/AbstractCommand.js
