+++ ozaria/site/components/cinematic/PageCinematicEditor/index.vue
