+++ test/app/lib/cinematic/DialogSystem.spec.js
@@ -1,5 +1,5 @@
 /* eslint-env jasmine */
-import { processText, wrapText } from '../../../../ozaria/lib/cinematic/dialog-system/dialogSystemHelper'
+import { processText, wrapText } from '../../../../ozaria/engine/cinematic/dialog-system/dialogSystemHelper'
 
 const dialogNode = text => ({ text })
