+++ test/app/models/Cinematic.spec.js
@@ -18,7 +18,7 @@ import {
   getSpeakingAnimationAction,
   getSetupMusic,
   getSoundEffects
-} from '../../../ozaria/schemas/selectors/cinematic'
+} from '../../../app/schemas/models/selectors/cinematic'
 
 /**
  * This data can be used to check that none of the selectors that match
